Fix lerna --scope option.
Fixes #567
This change allows to pass the `-- --scope ...` option when running `npm run test` or `npm run test-cov` from the project root (as stated in the README).
Before this change, the additional options configured in the global `package.json` would cause the `--scope` option to be forwarded to `jest` instead, causing an error.