Use paginated GraphQL API to list pipelines in build stream
Currently, the builds stream (contrary to the pipelines stream)
uses the REST API to get the list of pipelines to iterate over.
That method does not paginate, which means that when one has more
than 30 pipelines, builds will be missed.

We switch to the GraphQL API for listing pipelines, for which we paginate
properly.